Poll WIP (#231):
* mod_muc_poll refactoring
