+++ prosody-modules/mod_muc_poll/constants.lib.lua
@@ -1,0 +1,9 @@
+-- SPDX-FileCopyrightText: 2024 John Livingston <https://www.john-livingston.fr/>
+-- SPDX-License-Identifier: AGPL-3.0-only
+
+-- FIXME: create a XEP to standardize this, and remove the "x-".
+local xmlns_poll = "http://jabber.org/protocol/muc#x-poll";
+
+return {
+  xmlns_poll = xmlns_poll;
+};
